DRTII-656 Move DateLike and case classes to package
