+++ client/src/main/scala/drt/client/services/JSDateConversions.scala
@@ -2,7 +2,8 @@ package drt.client.services
 
 import drt.client.services.JSDateConversions.SDate.JSSDate
 import drt.shared.CrunchApi.MillisSinceEpoch
-import drt.shared.{LocalDate, MilliDate, SDateLike, UtcDate}
+import drt.shared.dates.{LocalDate, UtcDate}
+import drt.shared.{MilliDate, SDateLike}
 import moment._
 
 import scala.language.implicitConversions
