+++ server/src/main/scala/actors/migration/TerminalDayStaffMinutesMigrationActor.scala
@@ -5,8 +5,9 @@ import actors.acking.AckingReceiver.Ack
 import actors.{PostgresTables, RecoveryActorLike, Sizes}
 import akka.actor.Props
 import akka.persistence.{SaveSnapshotSuccess, SnapshotMetadata}
-import drt.shared.CrunchApi.{StaffMinute, MillisSinceEpoch}
-import drt.shared.{SDateLike, TM, UtcDate}
+import drt.shared.CrunchApi.{MillisSinceEpoch, StaffMinute}
+import drt.shared.dates.UtcDate
+import drt.shared.{SDateLike, TM}
 import org.slf4j.{Logger, LoggerFactory}
 import scalapb.GeneratedMessage
 import server.protobuf.messages.CrunchState.{StaffMinuteMessage, StaffMinutesMessage}
