+++ server/src/main/scala/actors/queues/DateRange.scala
@@ -3,7 +3,8 @@ package actors.queues
 import akka.NotUsed
 import akka.stream.scaladsl.Source
 import drt.shared.CrunchApi.MillisSinceEpoch
-import drt.shared.{DateLike, LocalDate, MilliTimes, SDateLike, UtcDate}
+import drt.shared.dates.{DateLike, LocalDate, UtcDate}
+import drt.shared.{MilliTimes, SDateLike}
 import org.checkerframework.checker.units.qual.m
 import services.SDate
