+++ server/src/main/scala/actors/queues/FlightsRouterActor.scala
@@ -15,6 +15,7 @@ import drt.shared.CrunchApi.MillisSinceEpoch
 import drt.shared.FlightsApi.{FlightsWithSplits, FlightsWithSplitsDiff}
 import drt.shared.Terminals.Terminal
 import drt.shared._
+import drt.shared.dates.UtcDate
 import services.SDate
 
 import scala.collection.immutable.NumericRange
