+++ server/src/main/scala/controllers/application/exports/WithDesksExport.scala
@@ -7,7 +7,8 @@ import controllers.Application
 import controllers.application.exports.CsvFileStreaming.{makeFileName, sourceToCsvResponse}
 import drt.shared.CrunchApi.MillisSinceEpoch
 import drt.shared.Terminals.Terminal
-import drt.shared.{DateLike, ErrorResponse, LocalDate, SDateLike}
+import drt.shared.dates.LocalDate
+import drt.shared.{ErrorResponse, SDateLike}
 import play.api.mvc.{Action, AnyContent}
 import services.SDate
 import services.exports.StreamingDesksExport
