+++ server/src/main/scala/services/graphstages/Crunch.scala
@@ -6,6 +6,7 @@ import drt.shared.Queues.Queue
 import drt.shared.Terminals.Terminal
 import drt.shared._
 import drt.shared.api.Arrival
+import drt.shared.dates.{DateLikeOrdering, UtcDate}
 import org.joda.time.{DateTime, DateTimeZone}
 import org.slf4j.{Logger, LoggerFactory}
 import services._
