+++ server/src/test/scala/services/simulations/SimulationParamsSpec.scala
@@ -5,6 +5,7 @@ import drt.shared.FlightsApi.FlightsWithSplits
 import drt.shared.PaxTypesAndQueues._
 import drt.shared.Terminals.Terminal
 import drt.shared._
+import drt.shared.dates.LocalDate
 import org.specs2.mutable.Specification
 
 class SimulationParamsSpec extends Specification {
